reduce python warnings

+++ src/modules/attention.py
@@ -175,7 +175,7 @@ class CrossAttention(nn.Module):
     """
 
     def __init__(
-        self, query_dim: int, context_dim: Optional[int] = None, heads: int = 8, dim_head: int = 64, dropout: int = 0.0
+        self, query_dim: int, context_dim: Optional[int] = None, heads: int = 8, dim_head: int = 64, dropout: float = 0.0
     ):
         super().__init__()
         inner_dim = dim_head * heads
